Battery life
Battery live uses a dynamic minimum SoC. When we reach the minimum,
discharging will no longer allowed until the SoC rises above the
minimum again. The SoC minimum is adjusted automaticallyto ensure
that the batteries will not be (almost) empty for a long time (for
example in winter). In times of little solar yield the minimum
SoC will be high, otherwise it will be low.