Exit process when a serialEvent is detected.
A serialEvent is emitted whenever a USB error or disconnect is
detected. In that case the process should be exited, so the
device (/dev/tty...) can be removed by the kernel if needed.

In earlier implementation of the serial port module in velib,
the application was stopped by calling pltExit. This method
is still in use, but is deprecated now. Using serialEvent is
now the preferred way to do this.