Set /Connected=1 on D-Bus service of PV inverter on phase 2
If a multi phase energy meter (eg. EM24) is used to measure
the grid on phase 1 and a PV inverter on phase 2, the
Connected path on the PV inverter D-Bus service would be set
to 0 instead of 1. This causes the GUI to show
'not connected' instead of the actual power in the device lisr
summary.

https://github.com/victronenergy/venus/issues/165