Check & adjust measurement mode of EM3xx/ET3xx grid meters on startup
New ET340 grid meters will have the measurement mode set to mode A
which means that the power reported by the meter will always be
positive, even if power is fed back to the grid.

This commit checks the measurement mode on startup, and change it
if the device is not in mode B yet.

The already happens with other grid meters (EM24 & EM1xx).