Disconnect a sensor after 5 errors (any error).
In the past it would only disconnect after 5 Timeout errors. Any other error
would not lead to disconnection, which means that if an Modbus Exception is
raised in response to reading an invalid register, which should never happen
for any supported device, cgwacs would not terminate cleanly. This causes it to
claim hardware it is not entitled to. We will therefore treat any 5 consecutive
errors as a reason to get out.