Use ClassAndVrmInstance from localsettings
Instead of counting from 30 upwards and using even numbers for
meters and odd for pv-inverters, allocate device instance numbers
using ClassAndVrmInstance functionality.

Settings moved to /Settings/Devices/<serial> so it lives in the same
place as the deviceInstance setting.

Settings for piggy-backing PV on L2 is stored as a separate device.

Serial number for piggy-backed PV has a _S suffix to dinstinguish it
from L1, otherwise the GUI cannot change the role.

This also solves the issue where two grid meters detected at the same
time was given the same DeviceInstance.

https://github.com/victronenergy/venus-private/issues/62