send data to dbus more often
With data bundled in less messages, we can afford to send it more
often. Flush the data out at the end of each modbus cycle, which
means roughly every 250ms, but only if it has changed.

Since the meters only refresh every 600ms or 750ms (or 2s, ET340)
the effect is that dbus updates every 750ms or so, with the most
recent data.