rename constants to distinguish meter types better
The Em300 series all have the registers in the same place, so rename
those to include the text Em300. The protocol as implemented at present
is for the ET340 (it has per-phase reverse energy counters at 0x60), so
the protocol constants are renamed to include the text Et340.

A future development will add a protocol for the Em-series, as well as
properly detect them.