add support for the older EM-300 meters
To date we only supported the ET-series, which has per-phase
reverse energy counters. Lately I discovered that some EM-series
meters also have these registers. We do however have a workaround
in place for other meters without these registers, so in the end
it's easier to just check for the existence of that register, and
fall back to the workaround if it does not exist.