Restore the update rate of EM24 meters to 1 second
In 6a32cc the update speed was increased for all the meters,
flushing the values to dbus as quickly as they change. A side
effect of this was that EM24 meters started to look like the
power level is jumping around a lot, at certain power levels.

Now we restore the previous operation to 1 second. The EM24
isn't fast enough for the fast ESS loop anyway, so there is
no reason to go too fast.

https://github.com/victronenergy/venus/issues/1017