em500: set measurement mode to b when piggybacking
Because the EM500 has no per-phase energy registers, we assume
the total reverse energy can be used for L1. This is because by
definition, PV inverters only make forward energy, so all reverse
energy must belong to the grid meter.

Unfortunately this doesn't work in Measurement Mode C. In this mode
the total power (L1+L2) is integrated to get the total energy, which
means the total reverse energy counter sees only the difference between
L1 and L2 when exporting.

This does not affect the other EM300-based meters, which are similar
to the EM500, because they only support measurement mode B.

To work around this, we set measurement mode B when piggybacking
is required, and measurement mode C otherwise.

Additionally: Always redo setup when piggybacking changes.