Change FirmwareVersion into a string
Firmware version was always published as an integer, but this
was never in the expected 24-bit format. Since it is from a third
party meter, it's also not really sensible to fudge it. Therefore
simply publish it as a string. User interfaces can simply display
it if it is a string (as we already do in dbus-modbus-client for
ABB meters), and format it by the Victron rules if it is an
integer.

https://github.com/victronenergy/gui-v2/issues/1301